Call parse, test refactors
